another round of fixes

+++ dropbox/safari_history.py
@@ -8,8 +8,10 @@ from collections import defaultdict
 from sitename import sitename
 
 ######  copy the history file to DBox
+
 SAFARI_PATH = os.environ['HOME'] + '/Library/Safari'
-copyfile(SAFARI_PATH + '/History.db', APP_PATH + 'safari_history.db')
+if os.path.isfile(SAFARI_PATH + '/History.plist'):
+	copyfile(SAFARI_PATH + '/History.db', APP_PATH + 'safari_history.db')
 ###### convert sqlite into json 
 con = sqlite3.connect(APP_PATH + 'safari_history.db')
 cursor = con.cursor()
